requirements.txt
pyyaml==6.0.2
crewai[tools]==0.114.0
plantuml==0.3.0
networkx==3.4.2
python-louvain==0.16
tree_sitter==0.24.0
tree-sitter-php==0.23.11
tree-sitter-java==0.23.5
kopyt==0.0.2
tree-sitter-python==0.23.6